Crate group_member_approval_smart_contract

Group Member Approval Smart Contract

This contract uses Cosmwasm’s provided architecture in conjunction with Provwasm to create a wasm smart contract that can be deployed to and interact with the Provenance Blockchain.

This solves an issue in the groups module: any admin of any group can add a member to a group without their knowledge. In order to create an ecosystem that respects group member involvement, this grants group members the ability to explicitly state that they intended to become a member of a group.

Modules

contract

The entrypoint for all external commands sent to the compiled wasm.

execute

All commands that are available when executing this contract. Contains all execution routes used by the contract file.

instantiate

Defines the contract instantiation process. Contains the functionality used in the contract file to instantiate an instance of the contract.

migrate

Defines the contract migration process. Contains the functionality used in the contract file to perform a migration to a new version.

query

Defines the contract query process. Contains the functionality used in the contract file to perform a contract query.

store

Contains all internal storage communication functionality. Contains all type definitions and functionality for interacting with contract internal storage.

types

Contains all declared structs for internal and external communication. Contains all types and base functionality used to construct the logic of the contract.

util

Contains helper functionality for contract code facilitation. Additional functionality that does not strictly belong to a category. Global helpers.