group_member_approval_smart_contract::contract
Function execute

pub fn execute(
    deps: DepsMut<'_>,
    env: Env,
    info: MessageInfo,
    msg: ExecuteMsg,
) -> Result<Response, ContractError>

The entry point used when an account initiates an execution process defined in the contract. This defines the primary purposes of the contract.

Parameters

deps A dependencies object provided by the cosmwasm framework. Allows access to useful resources like contract internal storage and a querier to retrieve blockchain objects.

env An environment object provided by the cosmwasm framework. Describes the contract’s details, as well as blockchain information at the time of the transaction.

info A message information object provided by the cosmwasm framework. Describes the sender of the instantiation message, as well as the funds provided as an amount during the transaction.

msg A custom execution message enum defined by this contract to allow multiple different processes to be defined for the singular execution route entry point allowed by the cosmwasm framework.