Module group_member_approval_smart_contract::execute::approve_group_membership

The core functionality of the contract. Allows a blockchain account to approve its membership for a given group id.

Functions

approve_group_membership

Invoked via the contract’s execution functionality. Adds an attribute to the signer that denotes that they affirm their membership in a Provenance Blockchain Group by setting an int value on the designated attribute equal to the group identifier. Note: Cosmwasm does not expose functionality to query the group module, so this route does not do any verification that the signer is actually a member of the approved group. However, consenting to either being or becoming a member of a group is simply an act of compliance. False claims made herein can be queried from the standard chain routes, which allows external consumers of this attribute to verify this statement after it has been made. The route does, however, validate that the account does not already have an attribute value affirming the existing group, preventing duplicate writes.