Module group_member_approval_smart_contract::util::prov_helpers

Utility functions that assist in Provenance Blockchain communication and data parsing.

Functions

get_all_attributes

Fetches all attributes for an address, handling paging as appropriate

get_group_id_attribute_values

Parses all group ids from the Provenance Blockchain Attributes provided by filtering for all values that match the given name and have an assigned int value.

get_group_id_attribute_values_paginated

Parses all group ids from the Provenance Blockchain Attributes provided by filtering for all values that match the given name and have an assigned int value.

msg_bind_name

Generates a name bind msg that will properly assign the given name value to a target address. Assumes the parent name is unrestricted or that the contract has access to bind a name to the parent name.