Function group_member_approval_smart_contract::util::prov_helpers::get_group_id_attribute_values_paginated

pub fn get_group_id_attribute_values_paginated<S: Into<String>>(
    attributes: Vec<QueryAttributesResponse>,
    name: S,
) -> Vec<Uint64>

Parses all group ids from the Provenance Blockchain Attributes provided by filtering for all values that match the given name and have an assigned int value.

Parameters

attributes Pages of Attributes fetched via a chain query.

name A Provenance Blockchain Name Module name used to write the attribute.